+++ src/models/gateway/events/channel_pins_update.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct ChannelPinsUpdate {
-    guild_id: Option<String>,
-    channel_id: String,
+    pub guild_id: Option<String>,
+    pub channel_id: String,
     //last_pin_timestamp:
 }