+++ src/models/gateway/events/guild_integrations_update.rs
@@ -2,5 +2,5 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct GuildIntegrationsUpdate {
-    guild_id: String,
+    pub guild_id: String,
 }