+++ src/models/gateway/events/guild_members_chunk.rs
@@ -1,10 +1,12 @@
 use crate::models::guild::GuildMember;
 use serde::{Deserialize, Serialize};
 
+// TODO: implement presences
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct GuildMembersChunk {
     pub guild_id: String,
     pub members: Vec<GuildMember>,
-    not_found: Option<Vec<String>>,
+    #[serde(default)]
+    pub not_found: Vec<String>,
     //presences: Vec<Presence>
 }