+++ src/models/gateway/events/guild_role_delete.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct GuildRoleDelete {
-    guild_id: String,
-    role_id: String,
+    pub guild_id: String,
+    pub role_id: String,
 }