+++ src/models/gateway/events/message_delete_bulk.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MessageDeleteBulk {
-    ids: Vec<String>,
-    channel_id: String,
-    guild_id: Option<String>
+    pub ids: Vec<String>,
+    pub channel_id: String,
+    pub guild_id: Option<String>,
 }