+++ src/models/gateway/events/message_reaction_add.rs
@@ -3,10 +3,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MessageReactionAdd {
-    user_id: String,    //
-    channel_id: String, //
-    message_id: String, //
-    guild_id: Option<String>,
-    member: Option<GuildMember>,
-    emoji: Emoji, //
+    pub user_id: String,    //
+    pub channel_id: String, //
+    pub message_id: String, //
+    pub guild_id: Option<String>,
+    pub member: Option<GuildMember>,
+    pub emoji: Emoji, //
 }