+++ src/models/gateway/events/message_reaction_remove.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MessageReactionRemove {
-    user_id: String,
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
-    emoji: Emoji,
+    pub user_id: String,
+    pub channel_id: String,
+    pub message_id: String,
+    pub guild_id: Option<String>,
+    pub emoji: Emoji,
 }