+++ src/models/gateway/events/message_reaction_remove_all.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct MessageReactionRemoveAll {
-    channel_id: String,
-    message_id: String,
-    guild_id: Option<String>,
+    pub channel_id: String,
+    pub message_id: String,
+    pub guild_id: Option<String>,
 }